Append samples from multiple SummarizedExperiment objects

Append samples from multiple SummarizedExperiment objects by column-binding them. This function is equivalent to cbind but provides a tidyverse-like interface.

# S3 method for class 'SummarizedExperiment'
append_samples(x, ..., .id = NULL)

Arguments

x

First SummarizedExperiment object to combine

...

Additional SummarizedExperiment objects to combine by samples

.id

Object identifier (currently not used)

Value

A combined SummarizedExperiment object

Examples

data(se)
append_samples(se, se)
#> Warning: tidySummarizedExperiment says: you have duplicated sample names, they will be made unique.
#> class: RangedSummarizedExperiment 
#> dim: 100 16 
#> metadata(2): '' ''
#> assays(1): counts
#> rownames(100): LRG_239 ENSG00000233845 ... ENSG00000196400
#>   ENSG00000115687
#> rowData names(1): bla
#> colnames(16): SRR1039508 SRR1039509 ... SRR1039520_1 SRR1039521_1
#> colData names(9): SampleName cell ... Sample BioSample