Efficiently bind multiple data frames by row and column

This is an efficient implementation of the common pattern of do.call(rbind, dfs) or do.call(cbind, dfs) for binding many data frames into one.

Arguments

...
Data frames to combine.

Each argument can either be a data frame, a list that could be a data frame, or a list of data frames.

When row-binding, columns are matched by name, and any missing columns will be filled with NA.

When column-binding, rows are matched by position, so all data frames must have the same number of rows. To match by value, not position, see mutate-joins.

.id
Data frame identifier.

When .id is supplied, a new column of identifiers is created to link each row to its original data frame. The labels are taken from the named arguments to bind_rows(). When a list of data frames is supplied, the labels are taken from the names of the list. If no names are found a numeric sequence is used instead.

add.cell.ids
from SummarizedExperiment 3.0 A character vector of length(x=c(x, y)). Appends the corresponding values to the start of each objects' cell names.

Value

bind_rows() and bind_cols() return the same type as the first input, either a data frame, tbl_df, or grouped_df.

Details

The output of bind_rows() will contain a column if that column appears in any of the inputs.

Examples

`%>%` <- magrittr::`%>%`
library(tibble)
tt <- tidySummarizedExperiment::pasilla 
bind_rows(tt, tt)

#> Warning: tidySummarizedExperiment says: you have duplicated sample names, they will be made unique.

#> # A SummarizedExperiment-tibble abstraction: 204,386 x 5
#> # Transcripts=14599 | Samples=14 | Assays=counts
#>    feature     sample counts condition type      
#>    <chr>       <chr>   <int> <chr>     <chr>     
#>  1 FBgn0000003 untrt1      0 untreated single_end
#>  2 FBgn0000008 untrt1     92 untreated single_end
#>  3 FBgn0000014 untrt1      5 untreated single_end
#>  4 FBgn0000015 untrt1      0 untreated single_end
#>  5 FBgn0000017 untrt1   4664 untreated single_end
#>  6 FBgn0000018 untrt1    583 untreated single_end
#>  7 FBgn0000022 untrt1      0 untreated single_end
#>  8 FBgn0000024 untrt1     10 untreated single_end
#>  9 FBgn0000028 untrt1      0 untreated single_end
#> 10 FBgn0000032 untrt1   1446 untreated single_end
#> # … with 40 more rows

num_rows <- nrow(tidySummarizedExperiment::as_tibble(tt))
tt %>% bind_cols(tibble(a=0, num_rows))

#> # A SummarizedExperiment-tibble abstraction: 102,193 x 7
#> # Transcripts=14599 | Samples=7 | Assays=counts
#>    feature     sample counts condition type           a num_rows
#>    <chr>       <chr>   <int> <chr>     <chr>      <dbl>    <int>
#>  1 FBgn0000003 untrt1      0 untreated single_end     0   102193
#>  2 FBgn0000008 untrt1     92 untreated single_end     0   102193
#>  3 FBgn0000014 untrt1      5 untreated single_end     0   102193
#>  4 FBgn0000015 untrt1      0 untreated single_end     0   102193
#>  5 FBgn0000017 untrt1   4664 untreated single_end     0   102193
#>  6 FBgn0000018 untrt1    583 untreated single_end     0   102193
#>  7 FBgn0000022 untrt1      0 untreated single_end     0   102193
#>  8 FBgn0000024 untrt1     10 untreated single_end     0   102193
#>  9 FBgn0000028 untrt1      0 untreated single_end     0   102193
#> 10 FBgn0000032 untrt1   1446 untreated single_end     0   102193
#> # … with 40 more rows