Split data frame by groups

group_split() works like base::split() but:

It uses the grouping structure from group_by() and therefore is subject to the data mask

It does not name the elements of the list based on the grouping as this only works well for a single character grouping variable. Instead, use group_keys() to access a data frame that defines the groups.

group_split() is primarily designed to work with grouped data frames. You can pass ... to group and split an ungrouped data frame, but this is generally not very useful as you want have easy access to the group metadata.

# S3 method for class 'SummarizedExperiment'
group_split(.tbl, ..., .keep = TRUE)

Arguments

.tbl

A tbl.

...

If .tbl is an ungrouped data frame, a grouping specification, forwarded to group_by().

.keep

Should the grouping columns be kept?

Value

A list of tibbles. Each tibble contains the rows of .tbl for the associated group and all the columns, including the grouping variables. Note that this returns a list_of which is slightly stricter than a simple list but is useful for representing lists where every element has the same type.

Lifecycle

group_split() is not stable because you can achieve very similar results by manipulating the nested column returned from tidyr::nest(.by =). That also retains the group keys all within a single data structure. group_split() may be deprecated in the future.

See also

Examples

data(pasilla, package = "tidySummarizedExperiment")
pasilla |> group_split(condition)
#> [[1]]
#> # A SummarizedExperiment-tibble abstraction: 58,396 × 5
#> # Features=14599 | Samples=4 | Assays=counts
#>    .feature    .sample counts type       condition
#>    <chr>       <chr>    <int> <chr>      <chr>    
#>  1 FBgn0000003 untrt1       0 single_end untreated
#>  2 FBgn0000008 untrt1      92 single_end untreated
#>  3 FBgn0000014 untrt1       5 single_end untreated
#>  4 FBgn0000015 untrt1       0 single_end untreated
#>  5 FBgn0000017 untrt1    4664 single_end untreated
#>  6 FBgn0000018 untrt1     583 single_end untreated
#>  7 FBgn0000022 untrt1       0 single_end untreated
#>  8 FBgn0000024 untrt1      10 single_end untreated
#>  9 FBgn0000028 untrt1       0 single_end untreated
#> 10 FBgn0000032 untrt1    1446 single_end untreated
#> # ℹ 40 more rows
#> 
#> [[2]]
#> # A SummarizedExperiment-tibble abstraction: 43,797 × 5
#> # Features=14599 | Samples=3 | Assays=counts
#>    .feature    .sample counts type       condition
#>    <chr>       <chr>    <int> <chr>      <chr>    
#>  1 FBgn0000003 trt1         0 single_end treated  
#>  2 FBgn0000008 trt1       140 single_end treated  
#>  3 FBgn0000014 trt1         4 single_end treated  
#>  4 FBgn0000015 trt1         1 single_end treated  
#>  5 FBgn0000017 trt1      6205 single_end treated  
#>  6 FBgn0000018 trt1       722 single_end treated  
#>  7 FBgn0000022 trt1         0 single_end treated  
#>  8 FBgn0000024 trt1        10 single_end treated  
#>  9 FBgn0000028 trt1         0 single_end treated  
#> 10 FBgn0000032 trt1      1698 single_end treated  
#> # ℹ 40 more rows
#> 
pasilla |> group_split(counts > 0)
#> [[1]]
#> # A SummarizedExperiment-tibble abstraction: 38,752 × 6
#> # Features=5536 | Samples=7 | Assays=counts
#>    .feature    .sample counts condition type       `counts > 0`
#>    <chr>       <chr>    <int> <chr>     <chr>      <lgl>       
#>  1 FBgn0000003 untrt1       0 untreated single_end FALSE       
#>  2 FBgn0000015 untrt1       0 untreated single_end FALSE       
#>  3 FBgn0000022 untrt1       0 untreated single_end FALSE       
#>  4 FBgn0000028 untrt1       0 untreated single_end FALSE       
#>  5 FBgn0000038 untrt1       0 untreated single_end FALSE       
#>  6 FBgn0000039 untrt1       0 untreated single_end FALSE       
#>  7 FBgn0000055 untrt1       0 untreated single_end FALSE       
#>  8 FBgn0000056 untrt1       0 untreated single_end FALSE       
#>  9 FBgn0000061 untrt1       0 untreated single_end FALSE       
#> 10 FBgn0000114 untrt1       0 untreated single_end FALSE       
#> # ℹ 40 more rows
#> 
#> [[2]]
#> # A SummarizedExperiment-tibble abstraction: 86,513 × 6
#> # Features=12359 | Samples=7 | Assays=counts
#>    .feature    .sample counts condition type       `counts > 0`
#>    <chr>       <chr>    <int> <chr>     <chr>      <lgl>       
#>  1 FBgn0000008 untrt1      92 untreated single_end TRUE        
#>  2 FBgn0000014 untrt1       5 untreated single_end TRUE        
#>  3 FBgn0000017 untrt1    4664 untreated single_end TRUE        
#>  4 FBgn0000018 untrt1     583 untreated single_end TRUE        
#>  5 FBgn0000024 untrt1      10 untreated single_end TRUE        
#>  6 FBgn0000032 untrt1    1446 untreated single_end TRUE        
#>  7 FBgn0000036 untrt1       2 untreated single_end TRUE        
#>  8 FBgn0000037 untrt1      15 untreated single_end TRUE        
#>  9 FBgn0000042 untrt1  101664 untreated single_end TRUE        
#> 10 FBgn0000043 untrt1   33402 untreated single_end TRUE        
#> # ℹ 40 more rows
#> 
pasilla |> group_split(condition, counts > 0)
#> [[1]]
#> # A SummarizedExperiment-tibble abstraction: 21,084 × 6
#> # Features=5271 | Samples=4 | Assays=counts
#>    .feature    .sample counts type       condition `counts > 0`
#>    <chr>       <chr>    <int> <chr>      <chr>     <lgl>       
#>  1 FBgn0000003 untrt1       0 single_end untreated FALSE       
#>  2 FBgn0000015 untrt1       0 single_end untreated FALSE       
#>  3 FBgn0000022 untrt1       0 single_end untreated FALSE       
#>  4 FBgn0000028 untrt1       0 single_end untreated FALSE       
#>  5 FBgn0000038 untrt1       0 single_end untreated FALSE       
#>  6 FBgn0000039 untrt1       0 single_end untreated FALSE       
#>  7 FBgn0000055 untrt1       0 single_end untreated FALSE       
#>  8 FBgn0000056 untrt1       0 single_end untreated FALSE       
#>  9 FBgn0000061 untrt1       0 single_end untreated FALSE       
#> 10 FBgn0000114 untrt1       0 single_end untreated FALSE       
#> # ℹ 40 more rows
#> 
#> [[2]]
#> # A SummarizedExperiment-tibble abstraction: 47,544 × 6
#> # Features=11886 | Samples=4 | Assays=counts
#>    .feature    .sample counts type       condition `counts > 0`
#>    <chr>       <chr>    <int> <chr>      <chr>     <lgl>       
#>  1 FBgn0000008 untrt1      92 single_end untreated TRUE        
#>  2 FBgn0000014 untrt1       5 single_end untreated TRUE        
#>  3 FBgn0000017 untrt1    4664 single_end untreated TRUE        
#>  4 FBgn0000018 untrt1     583 single_end untreated TRUE        
#>  5 FBgn0000024 untrt1      10 single_end untreated TRUE        
#>  6 FBgn0000032 untrt1    1446 single_end untreated TRUE        
#>  7 FBgn0000036 untrt1       2 single_end untreated TRUE        
#>  8 FBgn0000037 untrt1      15 single_end untreated TRUE        
#>  9 FBgn0000042 untrt1  101664 single_end untreated TRUE        
#> 10 FBgn0000043 untrt1   33402 single_end untreated TRUE        
#> # ℹ 40 more rows
#> 
#> [[3]]
#> # A SummarizedExperiment-tibble abstraction: 14,970 × 6
#> # Features=4990 | Samples=3 | Assays=counts
#>    .feature    .sample counts type       condition `counts > 0`
#>    <chr>       <chr>    <int> <chr>      <chr>     <lgl>       
#>  1 FBgn0000003 trt1         0 single_end treated   FALSE       
#>  2 FBgn0000022 trt1         0 single_end treated   FALSE       
#>  3 FBgn0000028 trt1         0 single_end treated   FALSE       
#>  4 FBgn0000038 trt1         0 single_end treated   FALSE       
#>  5 FBgn0000039 trt1         0 single_end treated   FALSE       
#>  6 FBgn0000056 trt1         0 single_end treated   FALSE       
#>  7 FBgn0000120 trt1         0 single_end treated   FALSE       
#>  8 FBgn0000157 trt1         0 single_end treated   FALSE       
#>  9 FBgn0000206 trt1         0 single_end treated   FALSE       
#> 10 FBgn0000216 trt1         0 single_end treated   FALSE       
#> # ℹ 40 more rows
#> 
#> [[4]]
#> # A SummarizedExperiment-tibble abstraction: 35,190 × 6
#> # Features=11730 | Samples=3 | Assays=counts
#>    .feature    .sample counts type       condition `counts > 0`
#>    <chr>       <chr>    <int> <chr>      <chr>     <lgl>       
#>  1 FBgn0000008 trt1       140 single_end treated   TRUE        
#>  2 FBgn0000014 trt1         4 single_end treated   TRUE        
#>  3 FBgn0000015 trt1         1 single_end treated   TRUE        
#>  4 FBgn0000017 trt1      6205 single_end treated   TRUE        
#>  5 FBgn0000018 trt1       722 single_end treated   TRUE        
#>  6 FBgn0000024 trt1        10 single_end treated   TRUE        
#>  7 FBgn0000032 trt1      1698 single_end treated   TRUE        
#>  8 FBgn0000036 trt1         1 single_end treated   TRUE        
#>  9 FBgn0000037 trt1        20 single_end treated   TRUE        
#> 10 FBgn0000042 trt1    127363 single_end treated   TRUE        
#> # ℹ 40 more rows
#>