Coerce lists, matrices, and more to data frames

`r lifecycle::badge("maturing")`

`as_tibble()` turns an existing object, such as a data frame or matrix, into a so-called tibble, a data frame with class [`tbl_df`]. This is in contrast with [tibble()], which builds a tibble from individual columns. `as_tibble()` is to [`tibble()`] as [base::as.data.frame()] is to [base::data.frame()].

`as_tibble()` is an S3 generic, with methods for: * [`data.frame`][base::data.frame()]: Thin wrapper around the `list` method that implements tibble's treatment of [rownames]. * [`matrix`][methods::matrix-class], [`poly`][stats::poly()], [`ts`][stats::ts()], [`table`][base::table()] * Default: Other inputs are first coerced with [base::as.data.frame()].

Arguments

x
A data frame, list, matrix, or other object that could reasonably be coerced to a tibble.

...
Unused, for extensibility.

rownames
How to treat existing row names of a data frame or matrix: * `NULL`: remove row names. This is the default. * `NA`: keep row names. * A string: the name of a new column. Existing rownames are transferred into this column and the `row.names` attribute is deleted. Read more in [rownames].

.name_repair
see tidyr

For compatibility only, do not use for new code.

Value

A tibble

Row names

The default behaviour is to silently remove row names.

New code should explicitly convert row names to a new column using the `rownames` argument.

For existing code that relies on the retention of row names, call `pkgconfig::set_config("tibble::rownames"=NA)` in your script or in your package's [.onLoad()] function.

Life cycle

Using `as_tibble()` for vectors is superseded as of version 3.0.0, prefer the more expressive maturing `as_tibble_row()` and `as_tibble_col()` variants for new code.

See also

Examples

tidySummarizedExperiment::pasilla %>%
    
    as_tibble()

#> # A tibble: 102,193 x 5
#>    transcript  sample counts condition type      
#>    <chr>       <chr>   <int> <chr>     <chr>     
#>  1 FBgn0000003 untrt1      0 untreated single_end
#>  2 FBgn0000008 untrt1     92 untreated single_end
#>  3 FBgn0000014 untrt1      5 untreated single_end
#>  4 FBgn0000015 untrt1      0 untreated single_end
#>  5 FBgn0000017 untrt1   4664 untreated single_end
#>  6 FBgn0000018 untrt1    583 untreated single_end
#>  7 FBgn0000022 untrt1      0 untreated single_end
#>  8 FBgn0000024 untrt1     10 untreated single_end
#>  9 FBgn0000028 untrt1      0 untreated single_end
#> 10 FBgn0000032 untrt1   1446 untreated single_end
#> # … with 102,183 more rows